requirements.txt
# Web Framework
Flask>=2.0.1

# HTTP Requests
requests>=2.26.0

# Environment Variables
python-dotenv>=0.19.0

# Data Processing
pandas>=1.3.0  # 用于处理 CSV 数据 (app/utils/data_loader.py)

# Logging
Logging  # 用于日志管理 (app/__init__.py)

# YAML Processing
PyYAML>=5.4.1  # 用于处理配置文件 (config.example.yml)

# Development Tools
pytest>=6.2.5  # 用于测试
black>=21.5b2  # 代码格式化
flake8>=3.9.2  # 代码检查